refactor: decompose GenerationQueue
